+++ guillotina_elasticsearch/commands/vacuum.py
@@ -4,7 +4,7 @@ from guillotina.component import get_adapter
 from guillotina.component import get_utility
 from guillotina.db import ROOT_ID
 from guillotina.db import TRASHED_ID
-from guillotina.db.reader import reader
+from guillotina.utils import get_object_by_uid
 from guillotina.interfaces import ICatalogUtility
 from guillotina.utils import get_containers
 from guillotina_elasticsearch.interfaces import IIndexManager
@@ -24,7 +24,10 @@ import logging
 logger = logging.getLogger('guillotina_elasticsearch_vacuum')
 
 GET_CONTAINERS = 'select zoid from {objects_table} where parent_id = $1'
-SELECT_BY_KEYS = '''SELECT zoid from {objects_table} where zoid = ANY($1)'''
+SELECT_BY_KEYS = f'''
+SELECT zoid from {{objects_table}}
+where zoid = ANY($1) AND parent_id != '{TRASHED_ID}'
+'''
 GET_CHILDREN_BY_PARENT = """
 SELECT zoid, parent_id, tid
 FROM {objects_table}
@@ -34,10 +37,10 @@ ORDER BY parent_id
 
 PAGE_SIZE = 1000
 
-GET_OBS_BY_TID = """
+GET_OBS_BY_TID = f"""
 SELECT zoid, parent_id, tid
-FROM {objects_table}
-WHERE of is NULL
+FROM {{objects_table}}
+WHERE of is NULL and parent_id != '{TRASHED_ID}'
 ORDER BY tid ASC, zoid ASC
 """
 
@@ -163,33 +166,15 @@ class Vacuum:
         if oid in self.cache:
             return self.cache[oid]
 
-        try:
-            result = self.txn._manager._hard_cache.get(oid, None)
-        except AttributeError:
-            from guillotina.db.transaction import HARD_CACHE  # noqa
-            result = HARD_CACHE.get(oid, None)
-        if result is None:
-            result = await self.txn._cache.get(oid=oid)
-
-        if result is None:
-            result = await self.tm._storage.load(self.txn, oid)
-
-        obj = reader(result)
-        obj.__txn__ = self.txn
-        if result['parent_id']:
-            obj.__parent__ = await self.get_object(result['parent_id'])
-        return obj
+        return await get_object_by_uid(oid)
 
     async def process_missing(self, oid, index_type='missing', folder=False):
         # need to fill in parents in order for indexing to work...
         logger.warning(f'Index {index_type} {oid}')
         try:
             obj = await self.get_object(oid)
-        except KeyError:
+        except (AttributeError, KeyError, TypeError, ModuleNotFoundError):
             logger.warning(f'Could not find {oid}')
-            return
-        except (AttributeError, TypeError, ModuleNotFoundError):
-            logger.warning(f'Could not find {oid}', exc_info=True)
             return  # object or parent of object was removed, ignore
         try:
             if folder: